Completed Problem 3
Adds solution to problem 3
Initially, tried to just find the highest prime,
but the dividing as we go made it possible. With the
first method, I left the comoputer for 3 hours and
it still hadn't completed.

What was leart? With big numbers, dividing it
is a really good idea and makes it a lot faster.