Added timing example for Problem 3
Add some functionality to time the difference between looking for the
primes starting at the top and starting at the bottom. At number 160 or
so, the tides change and the forwards method becomes exponentially
faster, slashing any multiplier. The backwards methods takes n times
longer with a multiplier of n. Exponentials takes too long, but I am
sure there would be a way to speed it up.

Also adds one or two small changes to the original Problem003.py