Solution for problem 4
Add solution for problem 4
Solution should work with any amount of digits but seems to not work
for the example given
Not fully working with 2 digits, but works with 3 (answer was correct).
100 * increase in complexity for each digit, so definitely not the
ideal solution and could be revised for speed. There must be some
clever way to not have to check all the factors e.g. checking 100 * 100
is obviously not going to give the biggest (but how to choose what is
obvious…)